requirements.txt
matplotlib==2.2.3
numpy==1.16.2
Pillow==5.2.0
torch==1.2.0
torchvision==0.4.0
